you get lost in the lows and the highs

+++ coders/time_variance_authority.c
@@ -6,23 +6,23 @@
 /*   By: brouane <brouane@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/29 12:33:32 by brouane           #+#    #+#             */
-/*   Updated: 2026/04/29 12:33:33 by brouane          ###   ########.fr       */
+/*   Updated: 2026/05/02 00:02:04 by brouane          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "codexion.h"
 
-unsigned long get_time_ms(void)
+long long get_time_ms(void)
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return ((unsigned long)tv.tv_sec * 1000) + (tv.tv_usec / 1000);
+    return ((long long)tv.tv_sec * 1000) + (tv.tv_usec / 1000);
 }
 
-void precise_sleep(unsigned long duration_ms, t_simulation *sim)
+void precise_sleep(long long duration_ms, t_simulation *sim)
 {
-    unsigned long start;
-    unsigned long elapsed;
+    long long start;
+    long long elapsed;
 
     start = get_time_ms();
     while (1)
